Use karma to preprocess individual files
We were using dist/bigscreen.js as the source file for testing before,
which isn't ideal because we don't get to take advantage of sourcemaps
that way. The one disadvantage to this is that we have to duplicate the
logic for how the JST template keys are created, and we're not testing
the dist JS at all, so if the JST template keys got borked in Grunt, we
wouldn't know about it from testing.